requirements.txt
asyncpraw
google-cloud-pubsub
python-dotenv
discord.py
psycopg2-binary